requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
Pillow==10.0.0
fpdf==1.7.2
moviepy==1.0.3